fixed gcc compilation bug

+++ src/middle_algorithm.cpp
@@ -102,7 +102,7 @@ int middle_algorithm(BMatrix stab_mat, BMatrix anticomms) {
                     anticomms_syndrome.set(i, sym_prod(anticomms.row(i), v));
 
                 // check for compatilble words of weight d-1
-                it = lower_bound(s1.begin(), s1.end(), std::make_pair(stab_syndome.vec[0], 0));
+                it = lower_bound(s1.begin(), s1.end(), std::make_pair(stab_syndome.vec[0], u64(0)));
                 if (it != s1.end() && it->first == stab_syndome.vec[0] && it->second != anticomms_syndrome.vec[0])
                     throw 2 * d - 1;
 
@@ -144,7 +144,7 @@ int middle_algorithm(BMatrix stab_mat, BMatrix anticomms) {
                     anticomms_syndrome.set(i, sym_prod(anticomms.row(i), v));
 
                 // check for compatilble words of weight d-1
-                it = lower_bound(s0.begin(), s0.end(), std::make_pair(stab_syndome.vec[0], 0));
+                it = lower_bound(s0.begin(), s0.end(), std::make_pair(stab_syndome.vec[0], u64(0)));
                 if (it != s0.end() && it->first == stab_syndome.vec[0] && it->second != anticomms_syndrome.vec[0])
                     throw 2 * d;
